Separate look_up_xtest to another script
