+++ contrib/checkbox-provider-ce-oem/bin/parse_optee_test.py
@@ -2,19 +2,22 @@
 
 import json
 import argparse
-import shlex
-import re
-from checkbox_support.snap_utils.snapd import Snapd
-from checkbox_support.snap_utils.system import get_gadget_snap
+from look_up_xtest import look_up_xtest
 
 
 def parse_json_file(filepath, filter=False, xtest=None):
     with open(filepath, 'r') as file:
         data = json.load(file)
     for test in data:
-        if (filter and test['suite'] == 'pkcs11') or \
-           (not filter and test['suite'] not in ['pkcs11']):
+        if check_suite(test['suite'], filter):
             print_test_info(test, xtest)
+
+
+def check_suite(suite, filter):
+    if filter:
+        return suite == 'pkcs11'
+    else:
+        return suite != 'pkcs11'
 
 
 def print_test_info(test, xtest):
@@ -24,40 +27,20 @@ def print_test_info(test, xtest):
     print("tool: {}\n".format(xtest))
 
 
-def look_up_xtest():
-    if Snapd().list("x-test"):
-        xtest = "x-test.xtest"
-    elif look_up_gadget() is not False:
-        xtest = look_up_gadget()
-    else:
-        xtest = None
-    return xtest
-
-
-def look_up_gadget():
-    gadget = get_gadget_snap()
-    xtest = False
-    snap = Snapd().list(gadget)
-    for app in snap["apps"]:
-        if app["name"] == "xtest":
-            xtest = ".".join([app["snap"], app["name"]])
-    return xtest
-
-
 def main():
     parser = argparse.ArgumentParser(description='Parse a JSON file.')
-    parser.add_argument('-f', '--filepath',
+    parser.add_argument('filepath',
                         help='The path of the file to parse.')
     parser.add_argument('-p', '--pkcs11',
                         help='To filter out PKCS11 for the suite.'
                         'field in JSON.',
                         action='store_true')
     args = parser.parse_args()
-    xtest = look_up_xtest()
-    if args.filepath is None:
-        print(xtest)
-    else:
-        parse_json_file(args.filepath, args.pkcs11, xtest)
+    try:
+        xtest = look_up_xtest()
+    except SystemExit:
+        xtest = None
+    parse_json_file(args.filepath, args.pkcs11, xtest)
 
 
 if __name__ == "__main__":
