+++ contrib/checkbox-provider-ce-oem/units/optee/jobs.pxu
@@ -21,13 +21,14 @@ command:
 
 id: ce-oem-optee/xtest-check
 category_id: optee
-_summary: Check if xtest is in the gedget snap.
+_summary: Check if xtest is in the system.
 plugin: shell
 user: root
 estimated_duration: 20.0
 command:
-    tool=$(parse_optee_test.py)
-    if [[ "$tool" != "None" ]]; then
+    tool=$(look_up_xtest.py)
+    exit_status=$?
+    if [[ "$exit_status" -eq 0 ]]; then
         echo "Info: Found xtest runnable $tool"
     else
         echo "Error: Not able to found xtest runnable tool"
@@ -43,7 +44,7 @@ user: root
 estimated_duration: 20.0
 depends: ce-oem-optee/xtest-check
 command:
-    tool=$(parse_optee_test.py)
+    tool=$(look_up_xtest.py)
     ta_path=""
     if [[ "$tool" == "x-test.xtest" ]]; then
         ta_path="$(find /var/snap -wholename */lib/optee_armtz)"
@@ -76,7 +77,7 @@ command:
     else
         filepath="$PLAINBOX_PROVIDER_DATA/optee-test.json"
     fi
-    parse_optee_test.py -f "$filepath"
+    parse_optee_test.py "$filepath"
 _summary: Collect the test cases support by OP-TEE test(xtest)
 
 id: ce-oem-optee-test-list-pkcs11
@@ -91,7 +92,7 @@ command:
     else
         filepath="$PLAINBOX_PROVIDER_DATA/optee-test.json"
     fi
-    parse_optee_test.py -f "$filepath" -p
+    parse_optee_test.py "$filepath" -p
 _summary: Collect the test cases related with PKCS11 support by OP-TEE test(xtest)
 
 unit: template
